backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
boto3

google-genai

pillow
